requirements.txt
streamlit
tensorflow
pillow
numpy
requests
opencv-python